Splitting UIUtility into appropriate extensions
