+++ Sources/Core/Extensions/CGFloat+Extensions.swift
@@ -18,29 +18,9 @@
 
 import UIKit
 
-enum UIUtility {
-  /**
-   Insets a CGSize with the insets in a UIEdgeInsets.
-   */
-  static func size(from size: CGSize, withInsets insets: UIEdgeInsets) -> CGSize {
-    var rect = CGRect.zero
-    rect.size = size
-    return rect.inset(by: insets).size
-  }
-
-  /**
-   Outsets a CGSize with the insets in a UIEdgeInsets.
-   */
-  static func size(_ size: CGSize, outsetBy insets: UIEdgeInsets) -> CGSize {
-    return CGSize(
-      width: insets.left + size.width + insets.right,
-      height: insets.top + size.height + insets.bottom
-    )
-  }
-
+extension CGFloat {
   /**
    Limits a CGFloat value, using the scale to limit to pixels (instead of points).
-
 
    The limitFunction is frequention floorf, ceilf or roundf.  If the scale is 2.0,
    you may get back values of *.5 to correspond to pixels.
@@ -53,36 +33,5 @@ enum UIUtility {
     ) -> CGFloat {
     let limitedFloat = limitFunction(Float(pointValue * screenScale))
     return CGFloat(limitedFloat) / screenScale
-  }
-
-  static func textSize(
-    text: String,
-    font: UIFont,
-    constrainingSize: CGSize,
-    lineBreakMode: NSLineBreakMode
-    ) -> CGSize {
-    guard !text.isEmpty else {
-      return .zero
-    }
-
-    let paragraphStyle = NSMutableParagraphStyle()
-    paragraphStyle.lineBreakMode = lineBreakMode
-    let attributes = [
-      NSAttributedString.Key.font: font,
-      NSAttributedString.Key.paragraphStyle: paragraphStyle
-    ]
-
-    let attributedString = NSAttributedString(string: text, attributes: attributes)
-    let boundingRectSize = attributedString.boundingRect(
-      with: constrainingSize,
-      options: NSStringDrawingOptions
-        .usesDeviceMetrics
-        .union(.usesLineFragmentOrigin)
-        .union(.usesFontLeading),
-      context: nil
-    ).size
-    let size = MathUtility.ceil(for: boundingRectSize)
-
-    return MathUtility.ceil(for: size)
   }
 }
